refactor(cli): consolidate CLI into unified Typer-based architecture
Major changes:
- Migrate all CLI modules from argparse to typer framework
- Move agentic_rl CLI from finetune/reinforcement/common/cli.py to cli/agentic_rl.py
- Add backward compatibility layer for legacy command format (e.g., fine_tunes.call)
- Support global -k/--api-key parameter extraction
- Auto-translate underscore parameters to dash format (--training_file_ids -> --training-file-ids)

Benefits:
- All 25 commands now use consistent typer interface
- Legacy users can still use old command format
- Cleaner code organization with all CLI in dashscope/cli/
- Better help messages and user experience with rich/typer
- Reduced code by ~350 lines through deduplication

Backward compatibility:
- 17 legacy command formats automatically translated
- 13 legacy parameter names supported
- Global api-key parameter works before command name
- Short parameters (-t, -m, -e, etc.) unchanged
